fix: switch network
